A80:Fix screen stretch when resize video window.
1.When resize video window smaller,we use G2D to
scale down video.and we use a continuous physical
memory as a buffer to copy and convert every image.

Data stream as follow:
YV12/IYUV -> NV12(buffer) -> (G2D scale down and convert) -> ARGB(framebuffer)

2.When part of video window out of screen,we calculate
the region of video in the screen, and set it show on screen.

Notice:YOU MUST CHMOD /dev/ion TO 766!
注意，必须修改/dev/ion的权限为766!